creat run.py file
